// src/app/models/EmpresaModel.ts
export class Empresa {
    id_empresa: number;
    nombre: string;
    razon_social : string;
    ruc: string;
    direccion: string;
    telefono: string;
    telefono_add: string;
    representante_legal: string;
    email_repre: string;
    cargo_repre: string;
    ci_repre: string;
    dni: string;
    pais: number;
    provincia: number;
    ciudad: number;
    localidad: number;
    email: string;
    site_web: string;
    logo_route: string;
    estado: string;
    usuario_creacion: string;
    fecha_creacion: Date;
    usuario_actualizacion: string;
    fecha_actualizacion: Date; 
}